+++ R/hmetad.R
@@ -26,7 +26,7 @@
 #'
 #' # neither function works with factors
 #' try(to_signed(factor(1)))
-#' tryCatch(to_unsigned(factor(1)), warning=function(w) w)
+#' tryCatch(to_unsigned(factor(1)), warning = function(w) w)
 #' @rdname signed
 #' @export
 to_signed <- function(x) ifelse(x, 1, -1)
@@ -73,18 +73,20 @@ to_unsigned <- function(x) as.numeric(x > 0)
 #'
 #' @export
 joint_response <- function(response, confidence, K) {
-  ifelse(response, confidence + K, K + 1 - confidence)
+  confidence <- as.integer(confidence)
+  as.integer(ifelse(response, confidence + K, K + 1 - confidence))
 }
 
 #' @rdname responses
 #' @export
 type1_response <- function(joint_response, K) {
-  as.integer(joint_response > K)
+  as.integer(as.integer(joint_response) > K)
 }
 
 #' @rdname responses
 #' @export
 type2_response <- function(joint_response, K) {
+  joint_response <- as.integer(joint_response)
   ifelse(joint_response > K,
     joint_response - K,
     K + 1 - joint_response
@@ -164,35 +166,152 @@ response_probabilities <- function(counts) {
   }
 }
 
+#' Infer the number of confidence levels in aggregated or non-aggregated data
+#'
+#' @param data A tibble containing the data to fit the model.
+#' * If `aggregate`==TRUE, `data` should have one row per observation with
+#'   columns `stimulus`, `response`, `confidence`, and any other variables in
+#'   `formula`
+#' * If `aggregate`==FALSE, it should be aggregated to have one row per cell of
+#'   the design matrix, with joint type 1/type 2 response counts in a matrix
+#'   column (see [aggregate_metad()]).
+#' @param aggregate If `TRUE`, automatically aggregate `data` by the variables
+#'   included in `formula` using [aggregate_metad()]. Otherwise, `data` should
+#'   already be aggregated.
+#' @param .stimulus The name of "stimulus" column
+#' @param .response The name of "response" column
+#' @param .confidence The name of "confidence" column
+#' @param .joint_response The name of "joint_response" column
+#' @param categorical If `FALSE` (default), use the multinomial likelihood over
+#'   aggregated data. If `TRUE`, use the categorical likelihood over individual
+#'   trials.
+#' @returns An integer describing the number of confidence levels in `data`
+#' @keywords internal
+#' @noRd
+infer_confidence_levels <- function(
+  data, aggregate = TRUE, .stimulus = "stimulus", .response = "response",
+  .confidence = "confidence", .joint_response = "joint_response",
+  .name = "N", categorical = FALSE
+) {
+  # check validity of stimulus column
+  if (aggregate || categorical) {
+    if (!(.stimulus %in% names(data))) {
+      stop(paste0('Data must have a column called "', .stimulus, '".'))
+    }
+    if (!all(as.integer(as.character(unique(pull(data, .stimulus)))) %in% c(0, 1))) {
+      stop(
+        'Stimulus column "', .stimulus, '" should only have values 0 or 1, but has values: ',
+        setdiff(as.integer(as.character(unique(pull(data, .stimulus)))), c(0, 1))
+      )
+    }
+  }
+
+  K <- NULL
+  if (categorical) {
+    # for categorical models, use joint_response column
+    K <- max(as.integer(pull(data, .name)))
+
+    if (K %% 2 != 0) {
+      stop(paste0(
+        "Joint response column `", .name,
+        "` must have an even number of levels, but has ", K, " levels."
+      ))
+    }
+
+    K <- as.integer(K / 2)
+  } else if (!aggregate) {
+    if (!(.name %in% names(data))) {
+      stop(paste0("Aggregated data must have a column called `", .name, "`."))
+    }
+
+    # if data is already aggregated, count number of columns
+    K <- ncol(pull(data, .name))
+    if (K %% 4 != 0) {
+      stop(paste0(
+        "Joint response count matrix `", .name,
+        "` in aggregated data must have a number of columns divisible by 4, but has ",
+        K, " columns."
+      ))
+    }
+    K <- as.integer(K / 4)
+  } else if (.confidence %in% names(data) && .response %in% names(data)) {
+    K <- max(as.integer(pull(data, .confidence)))
+
+    ## validate .confidence and .response columns
+    if (!all(as.integer(as.character(unique(pull(data, .response)))) %in% c(0, 1))) {
+      stop(
+        "Response column `", .response, "` should only have values 0 or 1, but has values: ",
+        as.integer(as.character(unique(pull(data, .response))))
+      )
+    } else if (!all(as.integer(unique(pull(data, .confidence))) %in% seq_len(K))) {
+      stop(
+        "Confidence column `", .confidence, "` should only have values 1:", K, ", but has values: ",
+        as.integer(as.character(unique(pull(data, .confidence))))
+      )
+    }
+  } else if (.joint_response %in% names(data)) {
+    K <- max(as.integer(pull(data, .joint_response)))
+
+    # validate joint response column
+    if ((K %% 2) != 0) {
+      stop(paste0(
+        "Joint response column `", .joint_response,
+        "` must have an even number of levels, but has ", K, " levels"
+      ))
+    } else if (!all(as.integer(unique(pull(data, .joint_response))) %in% seq_len(K))) {
+      stop(
+        "Joint response column `", .joint_response,
+        "` should only have values 1:", K, ", but has values: ",
+        as.integer(as.character(unique(pull(data, .joint_response))))
+      )
+    }
+
+    K <- as.integer(K / 2)
+  } else {
+    stop(paste0(
+      "Data must have a column called `", .joint_response,
+      "`, or two columns called `", .response, "` and `", .confidence, "`."
+    ))
+  }
+
+  # check validity of K
+  if (K <= 1) {
+    stop("Number of confidence levels (`K`) must be greater than 1.")
+  }
+
+  message(paste0("`hmetad` has inferred that there are K=", K, " confidence levels in the data. If this is incorrect, please set this manually using the argument `K=<K>`"))
+
+  K
+}
+
 #' Aggregate `data` by `response`, `confidence`, and other columns
 #'
 #' Counts number of rows in `data` with unique combinations values in the
 #' columns `response`, `confidence`, and any other columns in `...`.
 #'
 #' @param data The data frame to aggregate
-#' @param ... Grouping columns in `data`.
-#' These columns will be converted to factors.
+#' @param ... Grouping columns in `data`. These columns will be converted to
+#'   factors.
 #' @param .stimulus The name of "stimulus" column
 #' @param .response The name of "response" column
 #' @param .confidence The name of "confidence" column
 #' @param .joint_response The name of "joint_response" column
 #' @param .name The name of the resulting column containing trial counts
-#' @param K The number of confidence levels in `data`.
-#' If `NULL`, this is estimated from `data`.
-#'
-#' @details
-#' The data frame `data` must have one column with the name given by `.stimulus`.
-#' Additionally, it must have either:
+#' @param K The number of confidence levels in `data`. If `NULL`, this is
+#'   estimated from `data` using the maximum value of either the confidence
+#'   column or joint response column.
+#' @details The data frame `data` must have one column with the name given by
+#' `.stimulus`. Additionally, it must have either:
 #'   * Two columns with names given by `.response` and `.confidence`
 #'   * One column with the name given by `.joint_response`
 #'
 #' Finally, it must also have columns for any additional variables in `...`.
 #'
-#' @returns A tibble with one row per combination of the variables in `...`,
-#' and another column named by the value of `.response` containing trial counts.
-#' For \eqn{K} confidence levels, this will be an \eqn{N \times K*4} matrix, such that the
-#' columns represent (for stimulus \eqn{S}, type 1 response \eqn{R}, and
-#' type 2 response \eqn{C}):
+#' @returns A tibble with one row per combination of the variables in `...`, and
+#'   another column named by the value of `.response` containing trial counts.
+#'   For \eqn{K} confidence levels, this will be an \eqn{N \times K*4} matrix,
+#'   such that the columns represent (for stimulus \eqn{S}, type 1 response
+#'   \eqn{R}, and type 2 response \eqn{C}):
 #' \deqn{
 #'  [N_{S=0, R=0, C=K}, \ldots, N_{S=0, R=0, C=1}, \\
 #'  N_{S=0, R=1, C=1}, \ldots, N_{S=0, R=1, C=K}, \\
@@ -222,7 +341,7 @@ response_probabilities <- function(counts) {
 #'   ungroup() |>
 #'   mutate(joint_response = joint_response(
 #'     response, confidence,
-#'     n_distinct(confidence)
+#'     max(as.integer(confidence))
 #'   )) |>
 #'   select(-response, -confidence) |>
 #'   aggregate_metad()
@@ -243,61 +362,35 @@ aggregate_metad <- function(
 
     data <- tidyr::expand_grid(..., "{.stimulus}" := 0:1, "{.response}" := 0:1, "{.confidence}" := 1:K) |>
       mutate(
-        "{.joint_response}" := factor(joint_response(!!sym(.response), !!sym(.confidence), K)),
+        "{.joint_response}" := joint_response(!!sym(.response), !!sym(.confidence), K),
         n = 0
       ) |>
       select(-!!sym(.response), -!!sym(.confidence)) |>
       arrange(..., !!sym(.stimulus), !!sym(.joint_response))
   } else {
-    ## aggregate data if non-empty
+    # aggregate data if non-empty
 
-    ## fill in missing columns
+    # infer number of confidence levels
+    if (is.null(K)) {
+      K <- infer_confidence_levels(
+        data,
+        aggregate = TRUE, categorical = FALSE,
+        .stimulus = .stimulus, .response = .response, .confidence = .confidence,
+        .joint_response = .joint_response, .name = .name
+      )
+    }
+
+    # create joint response column if it doesn't exist
     if (!(.joint_response %in% names(data))) {
-      if (!(.response %in% names(data) && .confidence %in% names(data))) {
+      if (!(.confidence %in% names(data) && .response %in% names(data))) {
         stop(paste0(
-          'Data must have a column called "', .joint_response,
-          '", or two columns called "', .response, '" and "', .confidence, '".'
+          "Data must have a column called `", .joint_response,
+          "`, or two columns called `", .response, "` and `", .confidence, "`."
         ))
       }
 
-      # number of confidence levels
-      if (is.null(K)) {
-        K <- data |>
-          pull(!!sym(.confidence)) |>
-          n_distinct()
-      }
-
-      ## add joint_response column
       data <- data |>
-        mutate("{.joint_response}" := factor(
-          joint_response(!!sym(.response), !!sym(.confidence), K),
-          levels = 1:(2 * K)
-        ))
-    } else {
-      # number of confidence levels
-      if (is.null(K)) {
-        K <- as.integer((data |> pull(!!sym(.joint_response)) |> n_distinct()) / 2)
-      }
-
-      if (!(.response %in% names(data))) {
-        ## add response column
-        data <- data |>
-          mutate("{.response}" := factor(type1_response(!!sym(.joint_response), K)))
-      }
-
-      if (!(.confidence %in% names(data))) {
-        ## add confidence column
-        data <- data |>
-          mutate("{.confidence}" := factor(type2_response(!!sym(.joint_response), K)))
-      }
-    }
-
-    if (K <= 1) {
-      stop("Number of confidence levels (`K`) must be greater than 1.")
-    }
-
-    if (!(.stimulus %in% names(data))) {
-      stop(paste0('Data must have a column called "', .stimulus, '".'))
+        mutate("{.joint_response}" := joint_response(!!sym(.response), !!sym(.confidence), K))
     }
 
     if (!all(names(enquos(..., .named = TRUE)) %in% names(data))) {
@@ -311,23 +404,30 @@ aggregate_metad <- function(
       ))
     }
 
+    # aggregate data, filling in empty cells with zero
     data <- data |>
       ungroup() |>
-      mutate(
-        "{.stimulus}" := factor(!!sym(.stimulus)),
-        "{.joint_response}" := factor(!!sym(.joint_response), levels = 1:(2 * K)),
-        across(c(...), factor)
-      ) |>
-      group_by(...) |>
-      count(!!sym(.stimulus), !!sym(.joint_response), .drop = FALSE)
+      count(!!sym(.stimulus), !!sym(.joint_response), ...) |>
+      tidyr::complete(
+        data |>
+          ungroup() |>
+          tidyr::expand(
+            "{.stimulus}" := c(0L, 1L),
+            "{.joint_response}" := seq_len(2 * K),
+            ...
+          ),
+        fill = list(n = 0)
+      )
   }
 
+  # convert data to wide format (one row per cell in ...)
   data <- data |>
     tidyr::pivot_wider(
       names_from = all_of(c(.stimulus, .joint_response)),
       values_from = "n",
       names_prefix = glue::glue("{.name}_")
     ) |>
+    rowwise() |>
     mutate(
       "{.name}_0" := sum(c_across(starts_with(glue::glue("{.name}_0_"),
         ignore.case = FALSE
@@ -335,7 +435,18 @@ aggregate_metad <- function(
       "{.name}_1" := sum(c_across(starts_with(glue::glue("{.name}_1_"),
         ignore.case = FALSE
       )))
-    )
+    ) |>
+    ungroup()
+
+  # if data is not empty, remove any cells with no observations
+  if (!all(data[, glue::glue("{.name}_0")] == 0) ||
+    !all(data[, glue::glue("{.name}_1")] == 0)) {
+    data <- data |>
+      filter(
+        !!sym(glue::glue("{.name}_0")) > 0 |
+          !!sym(glue::glue("{.name}_1")) > 0
+      )
+  }
 
   # convert counts into a matrix-column
   tibble(
@@ -346,7 +457,8 @@ aggregate_metad <- function(
         glue::glue("{.name}_1")
       ))
     ),
-    "{.name}" := data |> ungroup() |>
+    "{.name}" := data |>
+      ungroup() |>
       select(matches(glue::glue("{.name}_([[:digit:]]+)_([[:digit:]]+)"),
         ignore.case = FALSE
       )) |>
@@ -377,8 +489,9 @@ aggregate_metad <- function(
 #' @param .response The name of "response" column
 #' @param .confidence The name of "confidence" column
 #' @param .joint_response The name of "joint_response" column
-#' @param K The number of confidence levels. By default, this is estimated from
-#'   the data.
+#' @param K The number of confidence levels in `data`. If `NULL`, this is
+#'   estimated from `data` using the maximum level of either the confidence
+#'   column or joint response column.
 #' @param distribution The noise distribution to use for the signal detection
 #'   model. By default, uses a normal distribution with a mean parameterized by
 #'   `dprime`.
@@ -391,12 +504,17 @@ aggregate_metad <- function(
 #' @param categorical If `FALSE` (default), use the multinomial likelihood over
 #'   aggregated data. If `TRUE`, use the categorical likelihood over individual
 #'   trials.
+#' @param logit If `TRUE` (default), use the logit parameterization of the
+#'   likelihood over the log joint response probabilities. If `FALSE`, use the
+#'   standard parameterization of the likelihood over the actual joint response
+#'   probabilities. In most cases, the logit parameterization should provide
+#'   more stable numerical computations, but the standard parameterization might
+#'   be preferable in some settings.
 #' @returns A `brmsfit` object containing the fitted model
 #' @details `fit_metad(formula, data, ...)` is approximately the same as
 #'   `brm(formula, data=aggregate_metad(data, ...), family=metad(...),
-#'   stanvars=stanvars_metad(...), ...)`. For some models,
-#'   it may often be easier to use the more explicit version
-#'   than using `fit_metad`.
+#'   stanvars=stanvars_metad(...), ...)`. For some models, it may often be
+#'   easier to use the more explicit version than using `fit_metad`.
 #' @examples
 #' # check which parameters the model has
 #' metad(3)
@@ -424,36 +542,44 @@ fit_metad <- function(formula, data, ..., aggregate = TRUE,
                       .confidence = "confidence", .joint_response = "joint_response",
                       K = NULL,
                       distribution = "normal", metac_absolute = TRUE, stanvars = NULL,
-                      categorical = FALSE) {
+                      categorical = FALSE, logit = TRUE) {
   data.aggregated <- data
 
   # ensure formula is a brmsformula
   if ("formula" %in% class(formula)) {
     formula <- bf(formula)
   } else if ("mvbrmsformula" %in% class(formula)) {
-    stop("Error: multivariate brms formulas not yet supported.")
+    stop("Multivariate brms formulas not yet supported.")
   }
 
   # determine response variable
   .name <- all.vars(formula$formula)[attr(terms(formula$formula), "response")]
-
-  # determine number of confidence levels
-  if (is.null(K)) {
-    if (aggregate && !categorical) {
-      K <- n_distinct(data$confidence)
-    } else if (!categorical) {
-      # if data is already aggregated, count number of columns
-      K <- ncol(pull(data, .name)) / 4
-    } else {
-      # for categorical models, use joint_response column
-      K <- n_distinct(pull(data, .name)) / 2
+  if (categorical) {
+    # use vint variable as stimulus column for categorical models
+    .stimulus <- stringr::str_extract(
+      deparse1(formula$formula), 
+      "^.* \\| vint\\((.*)\\) ~ .*$", 
+      group=1
+    )
+    
+    # ensure that the stimulus column exists in the model formula
+    if (is.na(.stimulus)) {
+      stop("Categorical models require a `vint` term in the response of the model formula. See `help('brmsformula')` for more information")
     }
   }
-
-  if (K <= 1) {
-    stop("Number of confidence levels (`K`) must be greater than 1.")
+  
+  # determine number of confidence levels
+  # (run whether K is specified or not to validate data columns)
+  K_inferred <- infer_confidence_levels(
+    data,
+    aggregate = aggregate, categorical = categorical,
+    .stimulus = .stimulus, .response = .response, .confidence = .confidence,
+    .joint_response = .joint_response, .name = .name
+  )
+  if (is.null(K)) {
+   K <- K_inferred
   }
-
+  
   ## convert the formula to use the metad family
   formula <- bf(formula,
     family = metad(K,
@@ -472,7 +598,7 @@ fit_metad <- function(formula, data, ..., aggregate = TRUE,
       data, !!!terms,
       .stimulus = .stimulus, .response = .response,
       .confidence = .confidence, .joint_response = .joint_response,
-      .name = .name,
+      .name = .name, K = K
     )
   }
 
@@ -480,7 +606,8 @@ fit_metad <- function(formula, data, ..., aggregate = TRUE,
   sv <- stanvars_metad(K,
     distribution = distribution,
     metac_absolute = metac_absolute,
-    categorical = categorical
+    categorical = categorical,
+    logit = logit
   )
   if (!is.null(stanvars)) {
     sv <- sv + stanvars